pyproject.toml
[project]
name = "benchmarking-therapy-ovarian-cancer"
version = "0.1.0"
description = ""
authors = [
    {name = "ge83cak",email = "berbuir.sa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "pypdf (>=5.7.0,<6.0.0)",
    "langchain-huggingface (>=0.3.0,<0.4.0)",
    "langchain-google-vertexai (>=2.0.27,<3.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
